* Fix a bug on the SqlHiloDao class that causes the cause the SqlHiLoGenerator class to output wrong values for Ids

+++ src/base/common/data/id/HiLoGenerator.cs
@@ -17,7 +17,7 @@ namespace Nohros.Data
       #endregion
 
       public long High { get; private set; }
-      public long MaxLow { get; private set; }
+      public int MaxLow { get; private set; }
     }
 
     readonly NextHighDelegate next_high_;
